Remove redundant features
Features removes:
- double
- var_generator
- multisig

Resolves #25
